Create settings_base
